+++ FCU-Wolverine/Menu.h
@@ -8,17 +8,16 @@
 class Menu {
 public:
 	Menu();
-	void StartMenu(int buzzer);
-	void MainDisplay(double voltValue, int bbrest, int bbrestChargeur, String modeValue, int bbtire);
-	void MenuFullBurst(int ROFFull, int ROFBurst, int QteBB, int sousMenuValue);
-	void MenuSniper(int timeBolt, bool greenLed, int sousMenuValue);
-	void MenuChargeur1(int chargeurCap, bool switchchargeur, bool buzzer, bool blocage, int sousMenuValue);
-	void MenuChargeur2(bool alarm, int alarmLowBB, bool chargingHandle, bool greenLight, int sousMenuValue);
-	void MenuSetting(int veille, bool verrou, bool LowBat, int sousMenuValue);
-	void MenuSetting2(int semi, int full, int dwel, int sousMenuValue);
+	void StartMenu(int8_t buzzer);
+	void MainDisplay(double voltValue, int8_t bbrest, int8_t bbrestChargeur, String modeValue, int8_t bbtire);
+	void MenuFullBurst(int8_t ROFFull, int8_t ROFBurst, int8_t QteBB, int8_t sousMenuValue);
+	void MenuSniper(int8_t timeBolt, bool greenLed, int8_t sousMenuValue);
+	void MenuChargeur1(int8_t chargeurCap, bool switchchargeur, bool buzzer, bool blocage, int8_t sousMenuValue);
+	void MenuChargeur2(bool alarm, int8_t alarmLowBB, bool chargingHandle, bool greenLight, int8_t sousMenuValue);
+	void MenuSetting(int8_t veille, bool verrou, bool LowBat, int8_t sousMenuValue);
+	void MenuSetting2(int8_t semi, int8_t full, int8_t dwel, int8_t sousMenuValue);
 	void MenuDever();
 	void MenuVer();
 };
 
 #endif
-
